fix separate seatSelection
