+++ website/src/routes/schedules/$scheduleId/form/-components/seat_selector/hooks/useSeatSelection.tsx
@@ -1,0 +1,12 @@
+import { useContext } from "react";
+import { SeatSelectionContext } from "../SeatSelectionContext";
+
+export const useSeatSelection = () => {
+  const context = useContext(SeatSelectionContext);
+  if (!context) {
+    throw new Error(
+      "useSeatSelection must be used within a SeatSelectionProvider"
+    );
+  }
+  return context;
+};
